refactor(harness): extract electron-mocks.ts once T24 lands the third helper
Session 3 brings the third mock-then-call helper online
(installOpenExternalMock for shell.openExternal, mirroring
installShowItemInFolderMock and installOpenDialogMock). Threshold from
the session prompt was met — pull the three install/get pairs out of
lib/claudeai.ts into a dedicated lib/electron-mocks.ts. The mocks are
generic Electron module patches (dialog, shell), not claude.ai-domain,
so the new home keeps claudeai.ts focused on AX-tree page-objects.

T17, T25 imports updated to point at the new module. T24 (added in the
follow-up commit) imports from electron-mocks.ts directly.
