+++ tools/test-harness/src/lib/claudeai.ts
@@ -109,112 +109,6 @@ export async function activateTab(
 	return { clicked: true };
 }
 
-// Replace dialog.showOpenDialog with a mock that records every call and
-// returns a canned result. Idempotent — re-installing within the same
-// Electron lifecycle is a no-op (guarded by
-// globalThis.__claudeAiDialogMockInstalled). Mirrors the shape of
-// QuickEntry.installInterceptor (quickentry.ts:86) so callers across
-// libs feel consistent.
-//
-// The first BrowserWindow positional arg is optional in Electron's API,
-// so the mock handles both `showOpenDialog(opts)` and
-// `showOpenDialog(window, opts)` shapes.
-export async function installOpenDialogMock(
-	inspector: InspectorClient,
-	cannedResult: { canceled: boolean; filePaths: string[] } = {
-		canceled: false,
-		filePaths: ['/tmp/claude-test-folder'],
-	},
-): Promise<void> {
-	const canned = JSON.stringify(cannedResult);
-	await inspector.evalInMain<null>(`
-		if (globalThis.__claudeAiDialogMockInstalled) return null;
-		const { dialog } = process.mainModule.require('electron');
-		globalThis.__claudeAiDialogCalls = [];
-		const original = dialog.showOpenDialog.bind(dialog);
-		dialog.showOpenDialog = async function(...args) {
-			const browserWindowArg = args[0]
-				&& typeof args[0] === 'object'
-				&& args[0].constructor
-				&& args[0].constructor.name === 'BrowserWindow';
-			const opts = browserWindowArg ? args[1] : args[0];
-			globalThis.__claudeAiDialogCalls.push({
-				ts: Date.now(),
-				nargs: args.length,
-				title: opts && opts.title,
-				properties: opts && opts.properties,
-			});
-			return ${canned};
-		};
-		void original;
-		globalThis.__claudeAiDialogMockInstalled = true;
-		return null;
-	`);
-}
-
-export interface OpenDialogCall {
-	ts: number;
-	nargs: number;
-	title?: string;
-	properties?: string[];
-}
-
-// Read the recorded call list. Returns [] if the mock was never
-// installed (rather than throwing) — that way pre-install reads in
-// retry loops are cheap.
-export async function getOpenDialogCalls(
-	inspector: InspectorClient,
-): Promise<OpenDialogCall[]> {
-	return await inspector.evalInMain<OpenDialogCall[]>(
-		`return globalThis.__claudeAiDialogCalls || []`,
-	);
-}
-
-// Replace electron.shell.showItemInFolder with a mock that records every
-// call without performing the underlying DBus FileManager1 / xdg-open
-// dispatch. Same idempotency-flag pattern as installOpenDialogMock.
-//
-// Why mock vs. invoke real: `showItemInFolder` is fire-and-forget on
-// Linux (returns void, no success signal). Invoking it for real opens
-// the host's actual file manager — fine in a click-chain test, but
-// disruptive when the assertion is just "the JS-level call is reachable
-// + accepts a path arg + the IPC layer terminates here". The mock keeps
-// the same assertion shape with no host side effect.
-export async function installShowItemInFolderMock(
-	inspector: InspectorClient,
-): Promise<void> {
-	await inspector.evalInMain<null>(`
-		if (globalThis.__claudeAiShowItemMockInstalled) return null;
-		const { shell } = process.mainModule.require('electron');
-		globalThis.__claudeAiShowItemCalls = [];
-		const original = shell.showItemInFolder.bind(shell);
-		shell.showItemInFolder = function(fullPath) {
-			globalThis.__claudeAiShowItemCalls.push({
-				ts: Date.now(),
-				path: typeof fullPath === 'string' ? fullPath : String(fullPath),
-			});
-			// Return undefined like the real method — callers don't
-			// inspect the return value.
-		};
-		void original;
-		globalThis.__claudeAiShowItemMockInstalled = true;
-		return null;
-	`);
-}
-
-export interface ShowItemInFolderCall {
-	ts: number;
-	path: string;
-}
-
-export async function getShowItemInFolderCalls(
-	inspector: InspectorClient,
-): Promise<ShowItemInFolderCall[]> {
-	return await inspector.evalInMain<ShowItemInFolderCall[]>(
-		`return globalThis.__claudeAiShowItemCalls || []`,
-	);
-}
-
 // A "compact pill" — the React component used by both the env pill and
 // the "Select folder…" pill. AX shape: `role: 'button'` with
 // `hasPopup === 'menu'`, scoped away from cowork sidebar row triggers
